Comprehensive refactoring of the full-stack application.
This commit includes a complete overhaul of the backend and frontend of the "Thirteen" card game application.

Backend:
- Refactored the monolithic `api.php` into a structured, router-based API with separate endpoint files for different concerns (auth, user, room, game).
- Consolidated all game logic into a stateless `Game` class in `game.php`, removing duplicated and unused code.
- Simplified the royalty calculation algorithm for better performance and readability.
- Added comments and PHPDoc blocks to improve code quality and maintainability.

Frontend:
- Replaced the inefficient short-polling mechanism with a more robust long-polling implementation for real-time game state updates.
- Updated the API client to work with the new backend structure.
- Added comments to the main components to improve code quality.
