+++ backend/api/index.php
@@ -1,52 +1,57 @@
-php
 <?php
+/**
+ * Main API entry point and router for the Thirteen card game.
+ *
+ * This script handles incoming API requests, sets up the environment,
+ * and routes the request to the appropriate endpoint handler.
+ */
+header('Access-Control-Allow-Origin: http://localhost:3000'); // Adjust for your frontend URL
+header('Access-Control-Allow-Credentials: true');
+header('Access-Control-Allow-Methods: GET, POST, OPTIONS');
+header('Access-Control-Allow-Headers: Content-Type, Authorization');
 
-// Basic entry point for the Dou Dizhu backend.
-header('Content-Type: application/json; charset=utf-8');
-header('Access-Control-Allow-Origin: https://ss.wenxiuxiu.eu.org');
-error_reporting(E_ALL); ini_set('display_errors', 1);
-session_start();
-
-// Placeholder for game logic and state management.
-// This will eventually include handling game rooms, player connections, card dealing,
-// game state updates, and win/loss conditions. This logic might be separated
-// into different files or classes.
-
-// Placeholder for API endpoints.
-// This will handle incoming requests from the frontend, such as:
-// - Joining a game room
-// - Sending player actions (e.g., playing cards, calling the landlord)
-// - Receiving game state updates
-// API logic will be handled in api.php
-
-// Include the API handler
-require_once __DIR__ . '/api.php'; // Make sure api.php is in the same directory or adjust the path
-require_once __DIR__ . '/game.php'; // Include game logic file (assuming it's needed)
-require_once __DIR__ . '/utils.php'; // Include utility file (assuming it's needed)
-require_once __DIR__ . '/config.php'; // Include config file (assuming it's needed)
-
-// Basic routing based on request path
-$request_uri = explode('/', trim($_SERVER['REQUEST_URI'], '/'));
-$api_endpoint = end($request_uri); // Get the last part of the URL
-
-
-// Handle API requests
-switch ($api_endpoint) {
-    case 'create_room':
-        handleCreateRoom();
-        break;
-    case 'join_room':
-        handleJoinRoom();
-        break;
-    case 'get_room_state':
-        handleGetRoomState();
-        break;
-    // TODO: Add other API endpoints here (e.g., /play_card, /call_landlord)
-    default:
-        // Handle invalid endpoint
-        http_response_code(404);
-        echo json_encode(['error' => 'Endpoint not found']);
-        break;
+if ($_SERVER['REQUEST_METHOD'] === 'OPTIONS') {
+    http_response_code(204);
+    exit();
 }
 
-?>
+session_start();
+
+header('Content-Type: application/json; charset=utf-8');
+error_reporting(E_ALL);
+ini_set('display_errors', 1);
+
+require_once 'db.php';
+require_once 'game.php';
+
+function send_json_error($code, $message, $details = null) {
+    http_response_code($code);
+    $response = ['success' => false, 'message' => $message];
+    if ($details) {
+        $response['details'] = $details;
+    }
+    echo json_encode($response);
+    exit();
+}
+
+$db = get_db();
+$endpoint = $_GET['endpoint'] ?? '';
+$request_method = $_SERVER['REQUEST_METHOD'];
+$data = json_decode(file_get_contents('php://input'), true) ?? [];
+
+$auth_endpoints = ['register', 'login', 'logout', 'check_session'];
+$user_endpoints = ['find_user', 'transfer_points'];
+$room_endpoints = ['matchmake', 'get_room_state'];
+$game_endpoints = ['start_game', 'submit_hand'];
+
+if (in_array($endpoint, $auth_endpoints)) {
+    require_once 'endpoints/auth.php';
+} elseif (in_array($endpoint, $user_endpoints)) {
+    require_once 'endpoints/user.php';
+} elseif (in_array($endpoint, $room_endpoints)) {
+    require_once 'endpoints/room.php';
+} elseif (in_array($endpoint, $game_endpoints)) {
+    require_once 'endpoints/game.php';
+} else {
+    send_json_error(404, 'Endpoint not found');
+}
